Make `mute-logging` part of `testutil`
To 1) ease the adoption and internal usage 2) avoid split-package issue when migrating to newer Java.
