+++ base/src/test/java/io/spine/validate/ValidateTest.java
@@ -36,7 +36,7 @@ import io.spine.test.type.Url;
 import io.spine.test.validate.Passport;
 import io.spine.test.validate.RequiredMsgFieldValue;
 import io.spine.testing.UtilityClassTest;
-import io.spine.testing.logging.MuteLogging;
+import io.spine.testing.logging.mute.MuteLogging;
 import io.spine.type.TypeName;
 import io.spine.validate.diags.ViolationText;
 import org.junit.jupiter.api.DisplayName;
