+++ pom.xml
@@ -12,7 +12,7 @@ all modules and does not describe the project structure per-subproject.
 
 <groupId>io.spine</groupId>
 <artifactId>spine-base</artifactId>
-<version>2.0.0-SNAPSHOT.47</version>
+<version>2.0.0-SNAPSHOT.48</version>
 
 <inceptionYear>2015</inceptionYear>
