+++ testlib/src/test/java/io/spine/testing/logging/mute/MuteLoggingTest.java
@@ -24,15 +24,27 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * This package contains the tests for the {@link io.spine.testing.logging.MuteLogging} JUnit
- * extension.
- */
+package io.spine.testing.logging.mute;
 
-@CheckReturnValue
-@ParametersAreNonnullByDefault
-package io.spine.testing.logging;
+import com.google.common.truth.ObjectArraySubject;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.extension.Extension;
 
-import com.google.errorprone.annotations.CheckReturnValue;
+import static com.google.common.truth.Truth.assertThat;
 
-import javax.annotation.ParametersAreNonnullByDefault;
+@DisplayName("@MuteLogging should")
+class MuteLoggingTest {
+
+    @Test
+    @DisplayName("be marked as an extension")
+    void annotated() {
+        Class<MuteLogging> annotation = MuteLogging.class;
+        ExtendWith extendsWith = annotation.getAnnotation(ExtendWith.class);
+        Class<? extends Extension>[] extensions = extendsWith.value();
+        ObjectArraySubject<Class<? extends Extension>> assertExtensions = assertThat(extensions);
+        assertExtensions.hasLength(1);
+        assertExtensions.asList().contains(MuteLoggingExtension.class);
+    }
+}
