+++ tools/javadoc-filter/src/test/java/io/spine/tools/javadoc/ExcludeInternalDocletTest.java
@@ -30,7 +30,7 @@ import com.sun.javadoc.ClassDoc;
 import com.sun.javadoc.MethodDoc;
 import com.sun.javadoc.RootDoc;
 import io.spine.testing.TestValues;
-import io.spine.testing.logging.MuteLogging;
+import io.spine.testing.logging.mute.MuteLogging;
 import io.spine.util.Exceptions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
